🛠️ fix(observability): stabilize declaration generation
Build declaration entries serially so each public subpath is self-contained and repeatable, eliminating the nondeterministic shared hash chunk.
